Refactor wily config into submodule to reduce cyclical imports
