+++ src/wily/commands/graph.py
@@ -11,7 +11,7 @@ import plotly.graph_objs as go
 import plotly.offline
 
 from wily import format_datetime, logger
-from wily.config import WilyConfig
+from wily.config.types import WilyConfig
 from wily.operators import resolve_metric, resolve_metric_as_tuple
 from wily.state import State
